Add augmented PiML dataset with model predictions.
This dataset is based on a simulated credit dataset created by PiML's maintainers and sent to SolasAI on 12/21/2022.  In addition to the features and variables in the original dataset, we have (1) formatted the data to make it easier to use with Solas, (2) added proportional estimates of protected class based on the PiML race variables, (3) renamed the race and gender variables so that they are intuitive (i.e., minority/majority and male/female), (4) created a train, validation, test split, and (5) added a prediction and offer column based on an XGBoost model that we created.